agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.51.0
mcp==1.12.4
httpx==0.25.2
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-json-logger==2.0.7
pymongo==4.6.0
mysql-connector-python==8.2.0
redis==5.0.1
websockets==12.0
PyJWT==2.8.0 